+++ dune/istl/solvers.hh
@@ -110,7 +110,7 @@ namespace Dune {
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("loopsolver", defaultIterativeSolverCreator<Dune::LoopSolver>());
@@ -181,7 +181,7 @@ namespace Dune {
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("gradientsolver", defaultIterativeSolverCreator<Dune::GradientSolver>());
@@ -218,7 +218,7 @@ namespace Dune {
                                 This is only supported for float and double field types.
     */
     CGSolver (LinearOperator<X,X>& op, Preconditioner<X,X>& prec,
-      scalar_real_type reduction, int maxit, int verbose, bool condition_estimate) : IterativeSolver<X,X>(op, prec, reduction, maxit, verbose),
+      scalar_real_type reduction, int maxit, int verbosity, bool condition_estimate) : IterativeSolver<X,X>(op, prec, reduction, maxit, verbosity),
       condition_estimate_(condition_estimate)
     {
       if (condition_estimate && !enableConditionEstimate) {
@@ -235,7 +235,7 @@ namespace Dune {
                                 This is only supported for float and double field types.
     */
     CGSolver (LinearOperator<X,X>& op, ScalarProduct<X>& sp, Preconditioner<X,X>& prec,
-      scalar_real_type reduction, int maxit, int verbose, bool condition_estimate) : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbose),
+      scalar_real_type reduction, int maxit, int verbosity, bool condition_estimate) : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbosity),
       condition_estimate_(condition_estimate)
     {
       if (condition_estimate && !(std::is_same<field_type,float>::value || std::is_same<field_type,double>::value)) {
@@ -253,8 +253,8 @@ namespace Dune {
     */
     CGSolver (std::shared_ptr<LinearOperator<X,X>> op, std::shared_ptr<ScalarProduct<X>> sp,
               std::shared_ptr<Preconditioner<X,X>> prec,
-              scalar_real_type reduction, int maxit, int verbose, bool condition_estimate)
-      : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbose),
+              scalar_real_type reduction, int maxit, int verbosity, bool condition_estimate)
+      : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbosity),
       condition_estimate_(condition_estimate)
     {
       if (condition_estimate && !(std::is_same<field_type,float>::value || std::is_same<field_type,double>::value)) {
@@ -380,7 +380,7 @@ namespace Dune {
 
           res.condition_estimate = max_eigv / min_eigv;
 
-          if (this->_verbose > 0) {
+          if (this->_verbosity > 0) {
             std::cout << "Min eigv estimate: " << Simd::io(min_eigv) << '\n';
             std::cout << "Max eigv estimate: " << Simd::io(max_eigv) << '\n';
             std::cout << "Condition estimate: "
@@ -406,7 +406,7 @@ namespace Dune {
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("cgsolver", defaultIterativeSolverCreator<Dune::CGSolver>());
@@ -591,7 +591,7 @@ namespace Dune {
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     template<class CountType>
     using Iteration = typename IterativeSolver<X,X>::template Iteration<CountType>;
   };
@@ -802,7 +802,7 @@ namespace Dune {
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("minressolver", defaultIterativeSolverCreator<Dune::MINRESSolver>());
@@ -845,8 +845,8 @@ namespace Dune {
        \copydoc LoopSolver::LoopSolver(L&,P&,double,int,int)
        \param restart number of GMRes cycles before restart
      */
-    RestartedGMResSolver (LinearOperator<X,Y>& op, Preconditioner<X,Y>& prec, scalar_real_type reduction, int restart, int maxit, int verbose) :
-      IterativeSolver<X,Y>::IterativeSolver(op,prec,reduction,maxit,verbose),
+    RestartedGMResSolver (LinearOperator<X,Y>& op, Preconditioner<X,Y>& prec, scalar_real_type reduction, int restart, int maxit, int verbosity) :
+      IterativeSolver<X,Y>::IterativeSolver(op,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -856,8 +856,8 @@ namespace Dune {
        \copydoc LoopSolver::LoopSolver(L&,S&,P&,double,int,int)
        \param restart number of GMRes cycles before restart
      */
-    RestartedGMResSolver (LinearOperator<X,Y>& op, ScalarProduct<X>& sp, Preconditioner<X,Y>& prec, scalar_real_type reduction, int restart, int maxit, int verbose) :
-      IterativeSolver<X,Y>::IterativeSolver(op,sp,prec,reduction,maxit,verbose),
+    RestartedGMResSolver (LinearOperator<X,Y>& op, ScalarProduct<X>& sp, Preconditioner<X,Y>& prec, scalar_real_type reduction, int restart, int maxit, int verbosity) :
+      IterativeSolver<X,Y>::IterativeSolver(op,sp,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -892,8 +892,8 @@ namespace Dune {
     RestartedGMResSolver (std::shared_ptr<LinearOperator<X,Y>> op,
                           std::shared_ptr<ScalarProduct<X>> sp,
                           std::shared_ptr<Preconditioner<X,Y>> prec,
-                          scalar_real_type reduction, int restart, int maxit, int verbose) :
-      IterativeSolver<X,Y>::IterativeSolver(op,sp,prec,reduction,maxit,verbose),
+                          scalar_real_type reduction, int restart, int maxit, int verbosity) :
+      IterativeSolver<X,Y>::IterativeSolver(op,sp,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -1014,7 +1014,7 @@ namespace Dune {
         // and if j < _maxit (do not restart on last iteration)
         if( res.converged != true && j < _maxit ) {
 
-          if(_verbose > 0)
+          if(_verbosity > 0)
             std::cout << "=== GMRes::restart" << std::endl;
           // get saved rhs
           b = b2;
@@ -1113,7 +1113,7 @@ namespace Dune {
     using IterativeSolver<X,Y>::_sp;
     using IterativeSolver<X,Y>::_reduction;
     using IterativeSolver<X,Y>::_maxit;
-    using IterativeSolver<X,Y>::_verbose;
+    using IterativeSolver<X,Y>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
     int _restart;
   };
@@ -1259,7 +1259,7 @@ namespace Dune {
         // and if still j < _maxit (do not restart on last iteration)
         if( res.converged != true && j < _maxit)
         {
-          if (_verbose > 0)
+          if (_verbosity > 0)
             std::cout << "=== fGMRes::restart" << std::endl;
           // get rhs
           v[0] = b;
@@ -1281,7 +1281,7 @@ private:
     using RestartedGMResSolver<X,Y>::_sp;
     using RestartedGMResSolver<X,Y>::_reduction;
     using RestartedGMResSolver<X,Y>::_maxit;
-    using RestartedGMResSolver<X,Y>::_verbose;
+    using RestartedGMResSolver<X,Y>::_verbosity;
     using RestartedGMResSolver<X,Y>::_restart;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
@@ -1326,8 +1326,8 @@ private:
        \copydoc LoopSolver::LoopSolver(L&,P&,double,int,int)
        \param restart number of GMRes cycles before restart
      */
-    GeneralizedPCGSolver (LinearOperator<X,X>& op, Preconditioner<X,X>& prec, scalar_real_type reduction, int maxit, int verbose, int restart = 10) :
-      IterativeSolver<X,X>::IterativeSolver(op,prec,reduction,maxit,verbose),
+    GeneralizedPCGSolver (LinearOperator<X,X>& op, Preconditioner<X,X>& prec, scalar_real_type reduction, int maxit, int verbosity, int restart = 10) :
+      IterativeSolver<X,X>::IterativeSolver(op,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -1338,8 +1338,8 @@ private:
        \param restart When to restart the construction of
        the Krylov search space.
      */
-    GeneralizedPCGSolver (LinearOperator<X,X>& op, ScalarProduct<X>& sp, Preconditioner<X,X>& prec, scalar_real_type reduction, int maxit, int verbose, int restart = 10) :
-      IterativeSolver<X,X>::IterativeSolver(op,sp,prec,reduction,maxit,verbose),
+    GeneralizedPCGSolver (LinearOperator<X,X>& op, ScalarProduct<X>& sp, Preconditioner<X,X>& prec, scalar_real_type reduction, int maxit, int verbosity, int restart = 10) :
+      IterativeSolver<X,X>::IterativeSolver(op,sp,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -1375,9 +1375,9 @@ private:
     GeneralizedPCGSolver (std::shared_ptr<LinearOperator<X,X>> op,
                           std::shared_ptr<ScalarProduct<X>> sp,
                           std::shared_ptr<Preconditioner<X,X>> prec,
-                          scalar_real_type reduction, int maxit, int verbose,
+                          scalar_real_type reduction, int maxit, int verbosity,
                           int restart = 10) :
-      IterativeSolver<X,X>::IterativeSolver(op,sp,prec,reduction,maxit,verbose),
+      IterativeSolver<X,X>::IterativeSolver(op,sp,prec,reduction,maxit,verbosity),
       _restart(restart)
     {}
 
@@ -1478,7 +1478,7 @@ private:
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
     int _restart;
   };
@@ -1515,7 +1515,7 @@ private:
       \param mmax is the maximal number of previous vectors which are orthogonalized against the new search direction.
     */
     RestartedFCGSolver (LinearOperator<X,X>& op, Preconditioner<X,X>& prec,
-                        scalar_real_type reduction, int maxit, int verbose, int mmax = 10) : IterativeSolver<X,X>(op, prec, reduction, maxit, verbose), _mmax(mmax)
+                        scalar_real_type reduction, int maxit, int verbosity, int mmax = 10) : IterativeSolver<X,X>(op, prec, reduction, maxit, verbosity), _mmax(mmax)
     {
     }
 
@@ -1525,7 +1525,7 @@ private:
       \param mmax is the maximal number of previous vectors which are orthogonalized against the new search direction.
     */
     RestartedFCGSolver (LinearOperator<X,X>& op, ScalarProduct<X>& sp, Preconditioner<X,X>& prec,
-                        scalar_real_type reduction, int maxit, int verbose, int mmax = 10) : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbose), _mmax(mmax)
+                        scalar_real_type reduction, int maxit, int verbosity, int mmax = 10) : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbosity), _mmax(mmax)
     {
     }
 
@@ -1537,9 +1537,9 @@ private:
     RestartedFCGSolver (std::shared_ptr<LinearOperator<X,X>> op,
                         std::shared_ptr<ScalarProduct<X>> sp,
                         std::shared_ptr<Preconditioner<X,X>> prec,
-                        scalar_real_type reduction, int maxit, int verbose,
+                        scalar_real_type reduction, int maxit, int verbosity,
                         int mmax = 10)
-      : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbose), _mmax(mmax)
+      : IterativeSolver<X,X>(op, sp, prec, reduction, maxit, verbosity), _mmax(mmax)
     {}
 
     /*!
@@ -1663,7 +1663,7 @@ private:
     using IterativeSolver<X,X>::_sp;
     using IterativeSolver<X,X>::_reduction;
     using IterativeSolver<X,X>::_maxit;
-    using IterativeSolver<X,X>::_verbose;
+    using IterativeSolver<X,X>::_verbosity;
     using Iteration = typename IterativeSolver<X,X>::template Iteration<unsigned int>;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("restartedfcgsolver", defaultIterativeSolverCreator<Dune::RestartedFCGSolver>());
@@ -1726,7 +1726,7 @@ private:
     using RestartedFCGSolver<X>::_sp;
     using RestartedFCGSolver<X>::_reduction;
     using RestartedFCGSolver<X>::_maxit;
-    using RestartedFCGSolver<X>::_verbose;
+    using RestartedFCGSolver<X>::_verbosity;
   };
   DUNE_REGISTER_ITERATIVE_SOLVER("completefcgsolver", defaultIterativeSolverCreator<Dune::CompleteFCGSolver>());
   /** @} end documentation */